Improve JavaDoc for decompositions
- Also removed unneeded casting for Complex128 objects.

+++ src/main/java/org/flag4j/arrays/sparse/CooCVector.java
@@ -375,7 +375,7 @@ public class CooCVector extends AbstractCooFieldVector<CooCVector, CVector, CooC
         double mag = 0;
 
         for(int i = 0, size=nnz; i < size; i++) {
-            Complex128 v = (Complex128) data[i];
+            Complex128 v = data[i];
             mag += (v.re*v.re + v.im*v.im);
         }
 
@@ -406,7 +406,7 @@ public class CooCVector extends AbstractCooFieldVector<CooCVector, CVector, CooC
 
         for (int i = 0; i < data.length; i++) {
             if (!data[i].isZero()) {
-                result = 31*result + ((Complex128) data[i]).hashCode();
+                result = 31*result + data[i].hashCode();
                 result = 31*result + Integer.hashCode(indices[i]);
             }
         }
@@ -438,7 +438,7 @@ public class CooCVector extends AbstractCooFieldVector<CooCVector, CVector, CooC
 
             // Get data up until the stopping point.
             for(int i = 0; i<stopIndex; i++) {
-                value = StringUtils.ValueOfRound((Complex128) data[i], precision);
+                value = StringUtils.ValueOfRound(data[i], precision);
                 width = padding + value.length();
                 value = centering ? StringUtils.center(value, width) : value;
                 result.append(String.format("%-" + width + "s", value));
@@ -452,7 +452,7 @@ public class CooCVector extends AbstractCooFieldVector<CooCVector, CVector, CooC
             }
 
             // Get last entry now
-            value = StringUtils.ValueOfRound((Complex128) data[size-1], precision);
+            value = StringUtils.ValueOfRound(data[size-1], precision);
             width = padding + value.length();
             value = centering ? StringUtils.center(value, width) : value;
             result.append(String.format("%-" + width + "s", value));
